style: rename middleware files to hooks
Middleware is an express term.  Hooks is the fastify equivalent.
