+++ src/app/core/access-checker/access-checker.controller.ts
@@ -4,7 +4,7 @@ import { FastifyInstance } from 'fastify';
 import accessCheckerService from './access-checker.service';
 import cacheEntryService from './cache/cache-entry.service';
 import { PagingQueryStringSchema, SearchBodySchema } from '../core.schemas';
-import { requireAdminAccess, requireLogin } from '../user/auth/auth.middleware';
+import { requireAdminAccess, requireLogin } from '../user/auth/auth.hooks';
 
 export default function (_fastify: FastifyInstance) {
 	const fastify = _fastify.withTypeProvider<JsonSchemaToTsProvider>();
