+++ src/app/core/user/eua/eua.hooks.ts
